+++ samples/Game.Unity.Agar/Server/App/Program.cs
@@ -1,3 +1,4 @@
+using Lakona.Game.Clustering.Postgres;
 using Lakona.Game.Server.Hosting;
 using Lakona.Game.Server.Observability;
 using Lakona.Rpc.Serializer.MemoryPack;
@@ -13,6 +14,7 @@ using OpenTelemetry.Trace;
 return await LakonaGameServer.RunAsync(args, static server => server
     .AddServices(static (services, configuration) =>
     {
+        services.AddLakonaPostgresClustering(configuration);
         var serviceName = ResolveServiceName("lakona-game-unity-agar");
         var serviceInstanceId = configuration["Lakona:Node:Id"];