+++ src/Lakona.Game.Server/Cluster/Membership/Table/MembershipSchemaException.cs
@@ -1,4 +1,4 @@
 namespace Lakona.Game.Cluster.Membership;
 
-internal sealed class MembershipSchemaException(string message, Exception innerException)
+public sealed class MembershipSchemaException(string message, Exception innerException)
     : InvalidOperationException(message, innerException);